[#7142] move WebhookPayloadSerializer from web/view/ to web/alarm/vo/sender/payload/
